requirements.txt
pandas
numpy
nltk
joblib
pillow
pytesseract
streamlit
scikit-learn
